Consolidate date Decorator to shared Core
